requirements.txt
blinker==1.4
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.28
protobuf==3.19.3
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.19.2
SQLAlchemy==1.4.31
Werkzeug==2.0.2
WTForms==3.0.1
yapf==0.32.0